+++ src/components/textarea/index.js
@@ -56,6 +56,15 @@ class Textarea extends React.Component {
   }
 
   /**
+   * Defines a custom input type for this component.
+   *
+   * @method inputType
+   */
+  get inputType() {
+    return 'textarea';
+  }
+
+  /**
    * Renders the component.
    *
    * @method render
@@ -65,7 +74,7 @@ class Textarea extends React.Component {
       <div className={ this.mainClasses }>
 
         { this.labelHTML }
-        <textarea { ...this.inputProps } />
+        { this.inputHTML }
         { this.validationHTML }
 
       </div>