+++ src/utils/decorators/input-validation/index.js
@@ -177,8 +177,8 @@ let InputValidation = (ComposedComponent) => class Component extends ComposedCom
     if (!this.state.errorMessage) { return null; }
 
     return [
-      <Icon key="0" type="error" className="base-input__icon base-input__icon--error" />,
-      <div key="1" className="base-input__message base-input__message--error">
+      <Icon key="0" type="error" className="common-input__icon common-input__icon--error" />,
+      <div key="1" className="common-input__message common-input__message--error">
         { this.state.errorMessage }
       </div>
     ];
@@ -193,7 +193,7 @@ let InputValidation = (ComposedComponent) => class Component extends ComposedCom
     let classes = super.mainClasses || "";
 
     if (!this.state.valid) {
-      classes += " base-input--error";
+      classes += " common-input--error";
     }
 
     return classes;
@@ -208,7 +208,7 @@ let InputValidation = (ComposedComponent) => class Component extends ComposedCom
     let classes = super.inputClasses || "";
 
     if (!this.state.valid) {
-      classes += " base-input__input--error";
+      classes += " common-input__input--error";
     }
 
     return classes;