+++ CHANGELOG.md
@@ -122,6 +122,7 @@ A new style and colour scheme has been applied to the carbon components library.
 * Updated Form Cancel Button to use History object.
 * Input components can now use custom classes.
 * Checkbox label now sits inline, and is reversable.
+* Added props on inputs for inline labels.
 * Added Rainbow chart component.
 
 # 0.0.1